removed existing hub page
