+++ src/components/index.ts
@@ -2,6 +2,7 @@
 export { default as AddApiPopup } from "./AddApiPopup";
 export { default as Analytics } from "./Analytics";
 export { default as APICard } from "./APICard";
+export { default as APILayout } from "./APILayout";
 export { default as DevAPICard } from "./DevAPICard";
 export { default as ApiPageLayout } from "./ApiPageLayout";
 export { default as ApiPageSidebar } from "./ApiPageSidebar";
@@ -23,17 +24,16 @@ export { default as Monetize } from "./Monetize";
 export { default as Navbar } from "./Navbar";
 export { default as PasswordStrengthMeter } from "./PasswordStrengthMeter";
 export { default as Pricing } from "./Pricing";
-export { default as ProxySecret } from "./ProxySecret"
+export { default as ProxySecret } from "./ProxySecret";
 export { default as Sidebar } from "./Sidebar";
 export { default as Schema } from "./Schema";
-export { default as SelectMulti } from "./SelectMulti"
+export { default as SelectMulti } from "./SelectMulti";
+export { default as SettingsPage } from "./SettingsPage";
 export { default as SplitView } from "./SplitView";
+export { default as Subscription } from "./Subscription";
 export { default as TabPanel } from "./TabPanel";
 export { default as Threat } from "./Threat";
 export { default as Transform } from "./Transform";
 export { default as UserProSideBar } from "./UserProSideBar";
 export { default as Version } from "./Version";
 export { default as Widget } from "./Widget";
-export { default as APILayout } from "./APILayout"
-export { default as Subscription } from "./Subscription"
-export { default as SettingsPage } from "./SettingsPage"
