+++ src/pages/APIHub.tsx
@@ -1,16 +1,18 @@
 import React from 'react'
 
 import { HomeNavbar, APIHubHeader, Footer } from '../sections'
+import APIGrid from "../components/APIGrid";
 
 
 const Pricing:React.FC = () => {
 
     return (
-        <>
-            <HomeNavbar />
-            <APIHubHeader />
-            <Footer />
-        </>
+    <div style={{background: "#FFF",width: "100%",minHeight: "100vh"}}>
+        <HomeNavbar />
+        <APIHubHeader />
+        <APIGrid />
+        <Footer />
+    </div>
     )
 }
